Closer to a working `README.md`.
Generating source snippits from the `edify.md`. Seems to be laying out
nicely in GitHub Markdown. Need to actually work though the rest of the
documentation and convert anything that was a sketch of an example into
a working example. The sort out these rules.

Checking in some crufty ideas, like `HEREDOC` function for testing, for
posterity.

See #114.